I finished the Loss_2Dsquare.py file by writing the X_f,X_ic & X_bc inside the loss_pinn_2D function.
Additionally, outcommented params. has been added
to the train.py script, that can be activated if we dont want to run
with the currently default params in loss_pinn_2D().

The params are; x0,y0,x_min,y_min,x_max,y_max,t_min, t_max, N
N is the vector describing the amount of grid points in respectivelly: the domain, the initial domain & the boundary of the domain.